uploadFile method moved to uploadFile.js
Removed UploadFile.js, moved the uploadFile method that recoder.js  and upload.js use to a new file uploadFile.js.
