+++ firebase/uploadFile.js
@@ -1,16 +1,6 @@
-import { useRef } from 'react'
-import firebase from '../../firebase/clientApp'
+import firebase from '../firebase/clientApp'
 
-const UploadFile = () => {
-    const inputFile = useRef(null)
-
-    const upload = () => {
-        // get file
-        var file = inputFile.current.files[0]
-        uploadFile(file)
-    }
-
-    const uploadFile = async ( file ) => {
+const uploadFile = async (file) => {
         var docRef = firebase.firestore().collection('videos').doc() // leave as .doc() for a random unique doc name to be assigned
         // create a storage ref to videos directory
         var storageRef = firebase.storage().ref('videos/' + docRef.id)
@@ -31,13 +21,6 @@ const UploadFile = () => {
             })
             .then(alert('Data was successfully sent to firestore'))
         })
-    }
-
-    return (
-        <div>
-            <input type="file" onChange={upload} ref={inputFile} />
-        </div>
-    )
 }
 
-export default UploadFile
+export default uploadFile
