+++ pages/upload.js
@@ -2,12 +2,19 @@ import Head from 'next/head'
 import Image from 'next/image'
 import styles from '../styles/Home.module.css'
 import React from 'react'
-import UploadFile from '../components/storage/UploadFile'
+import uploadFile from '../firebase/uploadFile'
+import { useRef } from 'react'
 
 export default function Upload() {
+    const inputFile = useRef(null)
+    const upload = () => {
+        // get file
+        var file = inputFile.current.files[0]
+        uploadFile(file)
+    }
     return (
         <div>
-            <UploadFile />
+            <input type="file" onChange={upload} ref={inputFile} />
         </div>
-    );
+    )
 }
